Split reaction error classification
