+++ utils/monitoring/dropReactionErrorClassification.ts
@@ -1,0 +1,171 @@
+const PROXY_REACTION_PERMISSION_DENIED_MESSAGE =
+  "Proxy doesn't have permission to add reactions";
+
+export type ReactionErrorKind =
+  | "network"
+  | "auth"
+  | "rate-limit"
+  | "server"
+  | "endpoint-contract";
+
+export function toErrorMessage(error: unknown): string {
+  if (error instanceof Error) {
+    return error.message;
+  }
+  if (typeof error === "string") {
+    return error;
+  }
+  if (typeof error === "object" && error) {
+    const typedError = error as {
+      message?: unknown;
+      error?: unknown;
+    };
+    if (typeof typedError.message === "string") {
+      return typedError.message;
+    }
+    if (typeof typedError.error === "string") {
+      return typedError.error;
+    }
+  }
+  return String(error);
+}
+
+export function toCaptureExceptionInput(error: unknown): Error {
+  if (error instanceof Error) {
+    return error;
+  }
+  return new Error(toErrorMessage(error));
+}
+
+function parseStatusCode(status: unknown): number | null {
+  if (typeof status === "number" && Number.isFinite(status)) {
+    return status;
+  }
+
+  if (typeof status === "string") {
+    const normalizedStatus = status.trim();
+    if (!/^\d+$/.test(normalizedStatus)) {
+      return null;
+    }
+    const parsed = Number.parseInt(normalizedStatus, 10);
+    return Number.isNaN(parsed) ? null : parsed;
+  }
+
+  return null;
+}
+
+function extractErrorStatusCode(error: unknown): number | null {
+  if (error === null || typeof error !== "object") {
+    return null;
+  }
+
+  const typedError = error as {
+    status?: unknown;
+    code?: unknown;
+    response?: {
+      status?: unknown;
+    };
+    cause?: {
+      status?: unknown;
+      code?: unknown;
+      response?: {
+        status?: unknown;
+      };
+    };
+  };
+
+  return (
+    parseStatusCode(typedError.status) ??
+    parseStatusCode(typedError.response?.status) ??
+    parseStatusCode(typedError.code) ??
+    parseStatusCode(typedError.cause?.status) ??
+    parseStatusCode(typedError.cause?.response?.status) ??
+    parseStatusCode(typedError.cause?.code)
+  );
+}
+
+function isNetworkError(error: unknown): boolean {
+  if (error instanceof TypeError) {
+    return true;
+  }
+
+  const normalizedMessage = toErrorMessage(error).toLowerCase();
+  return (
+    normalizedMessage.includes("failed to fetch") ||
+    normalizedMessage.includes("load failed") ||
+    normalizedMessage.includes("networkerror") ||
+    normalizedMessage.includes("network error") ||
+    normalizedMessage.includes("network request failed")
+  );
+}
+
+export function classifyReactionError(error: unknown): {
+  statusCode: number | null;
+  errorKind: ReactionErrorKind;
+} {
+  const statusCode = extractErrorStatusCode(error);
+
+  if (statusCode === 401 || statusCode === 403) {
+    return { statusCode, errorKind: "auth" };
+  }
+
+  if (statusCode === 429) {
+    return { statusCode, errorKind: "rate-limit" };
+  }
+
+  if (statusCode === 404 || statusCode === 405) {
+    return { statusCode, errorKind: "endpoint-contract" };
+  }
+
+  if (typeof statusCode === "number" && statusCode >= 500) {
+    return { statusCode, errorKind: "server" };
+  }
+
+  if (isNetworkError(error)) {
+    return { statusCode, errorKind: "network" };
+  }
+
+  return { statusCode, errorKind: "server" };
+}
+
+export function isExpectedStaleDropNotFoundError({
+  dropId,
+  endpoint,
+  errorMessage,
+  statusCode,
+}: {
+  readonly dropId: string;
+  readonly endpoint: string | null | undefined;
+  readonly errorMessage: string;
+  readonly statusCode: number | null;
+}): boolean {
+  // Intentionally exact: if the backend changes this stale-drop 404 shape,
+  // capture resumes so we can re-evaluate the contract.
+  return (
+    statusCode === 404 &&
+    endpoint === `drops/${dropId}/reaction` &&
+    errorMessage === `Drop ${dropId} not found`
+  );
+}
+
+export function isExpectedProxyReactionPermissionDeniedError({
+  dropId,
+  endpoint,
+  errorMessage,
+  method,
+  statusCode,
+}: {
+  readonly dropId: string;
+  readonly endpoint: string | null | undefined;
+  readonly errorMessage: string;
+  readonly method: string | null | undefined;
+  readonly statusCode: number | null;
+}): boolean {
+  // Intentionally exact: any different auth failure remains actionable.
+  return (
+    statusCode === 403 &&
+    endpoint === `drops/${dropId}/reaction` &&
+    method === "POST" &&
+    errorMessage === PROXY_REACTION_PERMISSION_DENIED_MESSAGE
+  );
+}
